requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# LangChain & AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
openai==1.10.0

# Notion Integration
notion-client==2.2.1

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Task Scheduling
apscheduler==3.10.4

# Environment & Config
python-dotenv==1.0.1

# Logging
python-json-logger==2.0.7

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0

# Utils
phonenumbers==8.13.27
